refactor(session): unify simulator event classification
